refactor(mesh): update page structure
